add contents

+++ src/main/java/com/unionpay/uplus/vo/UserVO.java
@@ -8,7 +8,7 @@ public class UserVO {
     /**
      * 用户id
      */
-    private long userId;
+    private int userId;
 
     /**
      * 用户 昵称
@@ -43,11 +43,11 @@ public class UserVO {
     public UserVO() {
     }
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
